Refactor duration form utils
